% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cotoha_as_df.r
\name{cotoha_as_df}
\alias{cotoha_as_df}
\title{Reshape response into data.frame}
\usage{
cotoha_as_df(response = response)
}
\arguments{
\item{response}{API response by \code{execute_cotoha_analysis} (json)}
}
\value{
Analysis result (data.frame)
}
\description{
Reshape response into data.frame
}
