% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_cotoha_analysis.r
\name{execute_cotoha_analysis}
\alias{execute_cotoha_analysis}
\title{Execute COTOHA analysis}
\usage{
execute_cotoha_analysis(
  access_token = "",
  API = "",
  sentence = "",
  document = "",
  s1 = "",
  s2 = "",
  optional = list()
)
}
\arguments{
\item{access_token}{Access token gotten by \code{get_cotoha_token} (character)}

\item{API}{API name (character): "parse", "ne", "coreference", "keyword", "similarity", "sentence_type", "user_attribute", "sentiment". See https://api.ce-cotoha.com/contents/reference/apireference.html.}

\item{sentence}{For "parse", "ne", "sentence_type", "sentiment" (character)}

\item{document}{For "coreference", "keyword", "user_attribute" (character)}

\item{s1}{For "similarity" (character)}

\item{s2}{For "similarity" (character)}

\item{optional}{Additional argument for each API (list). See https://api.ce-cotoha.com/contents/reference/apireference.html.}
}
\value{
HTTP response (json)
}
\description{
Execute COTOHA analysis
}
