% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cotoha_token.r
\name{get_cotoha_token}
\alias{get_cotoha_token}
\title{Get access token for COTOHA API}
\usage{
get_cotoha_token(config = "config.json", endpoint = "endpoint.json")
}
\arguments{
\item{config}{Your config file (json)}

\item{endpoint}{Your endpoint file (json)}
}
\value{
Access token (character)
}
\description{
Get access token for COTOHA API
}
