% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_cotoha_config.r
\name{make_cotoha_config}
\alias{make_cotoha_config}
\title{Write configulation files for COTOHA API}
\usage{
make_cotoha_config(clientId, clientSecret, tokenURL, baseURL)
}
\arguments{
\item{clientId}{Your client ID (character)}

\item{clientSecret}{Your client secret (character)}

\item{tokenURL}{Your access token publish url (character)}

\item{baseURL}{Your api base url (character)}
}
\value{
("config.json" & "endpoint.json" will be made on the current working directory.)
}
\description{
Write configulation files for COTOHA API
}
